moved componenets folder to src folder and added loading state for when articles are loading
