+++ src/App.jsx
@@ -2,8 +2,8 @@ import { useState } from 'react'
 import reactLogo from './assets/react.svg'
 import viteLogo from '/vite.svg'
 import './App.css'
-import HomePage from '../components/HomePage'
-import Header from '../components/Header'
+import HomePage from './components/HomePage'
+import Header from './components/Header'
 import { Routes, Route } from 'react-router-dom'
 
 function App() {
